upload jpg image

+++ poo/call/upload.php
@@ -6,21 +6,37 @@
         $verif = getimagesize($_FILES["image"]["tmp_name"]);
         $path = "../../userImages/" . $_SESSION["id"] . "_" . $_SESSION["lastname"] . "_" . $_SESSION["firstname"] . "/";
         $tmp = $_FILES["image"]["tmp_name"];
+        $extension = pathinfo($_FILES["image"]["name"], PATHINFO_EXTENSION);
+
         if($verif !== false)
         {
-            mkdir($path);
-            $newName = explode(".", $_FILES["image"]["name"]);
-            echo $new = "profil". '.' . end($newName);
-            move_uploaded_file($tmp, $path. $new);
+            if($extension == "jpg")
+            {
+                if(file_exists($path."profil.jpg"))
+                {
+                    unlink($path. "profil.jpg");
+                    $newName = explode(".", $_FILES["image"]["name"]);
+                    $new = "profil". '.' . end($newName);
+                    move_uploaded_file($tmp, $path. $new);
+                }
+                else
+                {
+                    mkdir($path);
+                    $newName = explode(".", $_FILES["image"]["name"]);
+                    echo $new = "profil". '.' . end($newName);
+                    move_uploaded_file($tmp, $path. $new);
+                }
+            }
+            else
+            {
+                echo "Le fichier doit etre en JPG.";
+                return 0;
+            }
         }
         else
         {
             echo "error";
             return 0;
         }
-    }
-    else
-    {
-        //echo "caca";
     }
 ?>
